+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2017-08-15
+### Added
+- Import/export for narratives
+
+### Fixed
+- QTextEdit + QGraphicsView transparency bug fix
+
 ## 2017-08-10
 ### Added
 - Drag and drop for narratives and slides